v- P5, P6 & P7 -v CompanyEditScreen: tempComp has been deleted, was redundant.
Product: New additional (default) constructor.
CalculatedProfitMargin() has been expanded to check if the sale & purchase price for the item inserted is 0 - This is due to issues with taking percentage of 0.

ProductEditScreen: Has been created. It displays a menu to create and/or edit a product. It can be opened through the ProductScreen.

ProductEditScreen includes:
EditProduct()
CreateProduct()
and Retry(), which the methods above use.